fix(client): publishing to both networks
often, publishing to both DHT and Relays network cause accidental CAS error,
as the DHT tries to publish after the relays have published to the same nodes,
with outdated CAS field. We now ignore that altogether.

On the other hand, we make sure that we await until publishing to both is completed.